# [Exporter.Geneva] Avoid allocation for serializing scopes