--- a/test/OpenTelemetry.Exporter.Geneva.Tests/GenevaLogExporterTests.cs
+++ b/test/OpenTelemetry.Exporter.Geneva.Tests/GenevaLogExporterTests.cs
@@ -357,8 +357,10 @@ public class GenevaLogExporterTests
         }
     }
 
-    [Fact]
-    public void SerializeILoggerScopes()
+    [Theory]
+    [InlineData(false)]
+    [InlineData(true)]
+    public void SerializeILoggerScopes(bool hasCustomFields)
     {
         string path = string.Empty;
         Socket senderSocket = null;
@@ -381,6 +383,11 @@ public class GenevaLogExporterTests
                 senderSocket.Listen(1);
             }
 
+            if (hasCustomFields)
+            {
+                exporterOptions.CustomFields = new string[] { "food", "Name", "Key1" };
+            }
+
             var exportedItems = new List<LogRecord>();
 
             using var loggerFactory = LoggerFactory.Create(builder => builder
@@ -391,6 +398,7 @@ public class GenevaLogExporterTests
                     options.AddGenevaLogExporter(options =>
                     {
                         options.ConnectionString = exporterOptions.ConnectionString;
+                        options.CustomFields = exporterOptions.CustomFields;
                     });
                 }));
 
@@ -408,8 +416,8 @@ public class GenevaLogExporterTests
 
             using (logger.BeginScope("MyOuterScope"))
             using (logger.BeginScope("MyInnerScope"))
-            using (logger.BeginScope("MyInnerInnerScope with {Name} and {Age} of custom", "John Doe", 35))
-            using (logger.BeginScope(new List<KeyValuePair<string, object>> { new KeyValuePair<string, object>("MyKey", "MyValue") }))
+            using (logger.BeginScope("MyInnerInnerScope with {Name} and {Age} of custom", "John Doe", 25))
+            using (logger.BeginScope(new List<KeyValuePair<string, object>> { new("Key1", "Value1"), new("Key2", "Value2") }))
             {
                 logger.LogInformation("Hello from {food} {price}.", "artichoke", 3.99);
             }
@@ -431,9 +439,39 @@ public class GenevaLogExporterTests
             var signal = (fluentdData as object[])[0] as string;
             var TimeStampAndMappings = ((fluentdData as object[])[1] as object[])[0];
             var mapping = (TimeStampAndMappings as object[])[1] as Dictionary<object, object>;
-            Assert.Equal("John Doe", mapping["Name"]);
-            Assert.Equal((byte)35, mapping["Age"]);
-            Assert.Equal("MyValue", mapping["MyKey"]);
+
+            if (hasCustomFields)
+            {
+                var envProperties = mapping["env_properties"] as Dictionary<object, object>;
+
+                // Custom Fields
+                Assert.Equal("artichoke", mapping["food"]);
+                Assert.Equal("John Doe", mapping["Name"]);
+                Assert.Equal("Value1", mapping["Key1"]);
+
+                Assert.False(mapping.ContainsKey("MyOuterScope"));
+                Assert.False(mapping.ContainsKey("MyInnerScope"));
+
+                // env_properties
+                Assert.True(Equals(envProperties["price"], 3.99));
+                Assert.Equal((byte)25, envProperties["Age"]);
+                Assert.Equal("Value2", envProperties["Key2"]);
+
+                Assert.False(envProperties.ContainsKey("MyOuterScope"));
+                Assert.False(envProperties.ContainsKey("MyInnerScope"));
+            }
+            else
+            {
+                Assert.Equal("artichoke", mapping["food"]);
+                Assert.True(Equals(mapping["price"], 3.99));
+                Assert.Equal("John Doe", mapping["Name"]);
+                Assert.Equal((byte)25, mapping["Age"]);
+                Assert.Equal("Value1", mapping["Key1"]);
+                Assert.Equal("Value2", mapping["Key2"]);
+
+                Assert.False(mapping.ContainsKey("MyOuterScope"));
+                Assert.False(mapping.ContainsKey("MyInnerScope"));
+            }
 
             // Check other fields
             Assert.Single(exportedItems);